feat(item): adjust item creation with new status model
